==> notebooks/1_extracao_faixa_temporal_melhor_distribuicao_valores.ipynb <==
# Etapa 1: Identificação da faixa temporal com melhor distribuição de valores

## 1. Importando libs

!pip3 install pandas
!pip3 install numpy

import pandas as pd
import zipfile

## 2. Carregando o dataframe com todos os dados de todos os países e da faixa temporal de anos.

zf = zipfile.ZipFile('../data/Edstats_csv.zip')
text_files = zf.infolist()
df_todos_paises = pd.DataFrame()
for csv_file in text_files:        
    if csv_file.filename == 'EdStatsData.csv':
        print("Abrindo o arquivo",csv_file.filename)
        df_todos_paises = pd.read_csv(zf.open(csv_file.filename))
        break

if df_todos_paises.empty:
    exit("Erro ao ler o arquivo 'EdStatsData.csv' que fica dentro do 'Edstats_csv.zip' na pasta 'data'.")
else:
    print("Arquivo 'EdStatsData.csv' carregado com sucesso.")

## 3. Separando dados referentes ao Brasil.
O Brasil será o país selecionado, no trabalho de conclusão de curso, como referência para comparações, análises e previsões. Logo, é pertinente que uma faixa temporal seja escolhida de acordo com os seus dados.

df_brasil = df_todos_paises.loc[df_todos_paises['Country Name'].isin(['Brazil'])]
df_brasil.index = range(df_brasil.shape[0])

## 4. Selecionar as linhas referente aos indicadores relevantes ao tema e objetivos do trabalho de conclusão de curso.
Previamente, de acordo com o tema e os objetivos do trabalho de conclusão de curso, foram selecionados indicadores para responder questionamentos e realizar análise e previsões. Sendo esses considerados os indicadores relevantes.

df_indicadores_relevantes = pd.read_csv('../data/codigos_indicadores_relevantes.csv')

df_indicadores_relevantes.columns = ['Indicator Code']

df_faixa_temporal_indicadores_relevantes_brasil = df_brasil.loc[df_brasil['Indicator Code'].isin(df_indicadores_relevantes['Indicator Code'])]

df_faixa_temporal_indicadores_relevantes_brasil = df_faixa_temporal_indicadores_relevantes_brasil.drop(['Country Name', 'Country Code', 'Indicator Name', 'Indicator Code'], axis=1)

df_faixa_temporal_indicadores_relevantes_brasil.index = range(df_faixa_temporal_indicadores_relevantes_brasil.shape[0])

df_faixa_temporal_indicadores_relevantes_brasil.fillna(0, inplace = True)

df_anos_com_dados_vazios = df_faixa_temporal_indicadores_relevantes_brasil.loc[:, df_faixa_temporal_indicadores_relevantes_brasil.isin([0]).mean() >= .6]

df_intervalo_anos_preenchidos = df_faixa_temporal_indicadores_relevantes_brasil.drop(list(df_anos_com_dados_vazios.columns), axis=1)

if (df_intervalo_anos_preenchidos['2013'].mean().round(2) > df_intervalo_anos_preenchidos['2014'].mean().round(2)):
       
    anos = pd.Series(list(df_intervalo_anos_preenchidos))
    anos.name = 'Year'
    anos.drop(index = [16]).to_csv('../data/lista_anos.csv', index = False)
    
    print('A coluna de testes será a de 2013 e o intervalo de anos será entre 1998 a 2012')
    
else:
    print('A coluna de testes será a de 2014 e o intervalo de anos será entre 1998 a 2013')
