==> notebooks/1_extracao_faixa_temporal_melhor_distribuicao_valores_aids.ipynb <==
# Etapa 1: Identificação da faixa temporal com melhor distribuição de valores

## 1. Importando libs

!pip3 install pandas
!pip3 install numpy

import pandas as pd
import zipfile

## 2. Carregando o dataframe com todos os dados de todos os países e da faixa temporal de anos.

zf = zipfile.ZipFile('../data/Edstats_csv.zip')
text_files = zf.infolist()
df_todos_paises = pd.DataFrame()
for csv_file in text_files:        
    if csv_file.filename == 'EdStatsData.csv':
        print("Abrindo o arquivo",csv_file.filename)
        df_todos_paises = pd.read_csv(zf.open(csv_file.filename))
        break

if df_todos_paises.empty:
    exit("Erro ao ler o arquivo 'EdStatsData.csv' que fica dentro do 'Edstats_csv.zip' na pasta 'data'.")
else:
    print("Arquivo 'EdStatsData.csv' carregado com sucesso.")

## 3. Selecionar as linhas referentes aos indicadores relevantes ao tema relacionado à mensuração de pessoas com AIDS.
Previamente, de acordo com o tema e os objetivos do trabalho de conclusão de curso, foram selecionados indicadores socioeconômicos para realizar análises e previsões. Sendo esses considerados os indicadores relevantes.

### 3.1. Carregar códigos dos indicadores relacionados à AIDS.

df_indicadores_relevantes = pd.read_csv('../data/codigos_indicadores_relevantes_aids.csv')

df_indicadores_relevantes.columns = ['Indicator Code']

### 3.2. Remover as colunas que não serão utilizadas na mensuração da faixa temporal.

df_faixa_temporal_indicadores_relevantes = df_todos_paises.loc[df_todos_paises['Indicator Code'].isin(df_indicadores_relevantes['Indicator Code'])]

df_faixa_temporal_indicadores_relevantes = df_faixa_temporal_indicadores_relevantes.drop(['Country Name', 'Country Code', 'Indicator Name', 'Indicator Code'], axis=1)
df_faixa_temporal_indicadores_relevantes.index = range(df_faixa_temporal_indicadores_relevantes.shape[0])

df_faixa_temporal_indicadores_relevantes.fillna(0, inplace = True)
df_faixa_temporal_indicadores_relevantes.head()

### 3.3. Identificar a faixa temporal, relacionada aos indicadores relevantes ao tema, de acordo com o preenchimento dos dados no dataset "EdStatsData.csv".

df_anos_com_dados_vazios = df_faixa_temporal_indicadores_relevantes.loc[:, df_faixa_temporal_indicadores_relevantes.isin([0]).mean() >= 0.7]

df_intervalo_anos_preenchidos = df_faixa_temporal_indicadores_relevantes.drop(list(df_anos_com_dados_vazios.columns), axis=1)

df_intervalo_anos_preenchidos.head()

### 3.4. Identificar a faixa temporal, relacionada aos indicadores relevantes ao tema, de acordo com preenchimento dos dados no dataset "Epidemic transition metrics_Trend of new HIV infections.csv".

df_infeccoes_aids = pd.read_csv('../data/Epidemic transition metrics_Trend of new HIV infections.csv')

df_infeccoes_aids.head()

### 3.5. Exportar a faixa temporal em "lista_anos_aids.csv" com a interseção do período de anos dos dois datasets mencionados acima.
O ano que iniciará essa faixa de tempo será o de 2010, vide o resultado da célula anterior que exibe algumas linhas do dataset "Epidemic transition metrics_Trend of new HIV infections.csv" e percebe-se que este inicia-se em 2010.

Antes de escolher o ano final dessa faixa temporal, é preciso analisar a qualidade dos dados de 2015 e 2016 (já que os dados da coluna de 2017 estão em branco) do dataset "EdStatsData.csv". Com esse valor, será possível determinar que a lista de anos a ser utilizada nas próximas etapas será a de 2010 até 2016.

if (df_intervalo_anos_preenchidos['2015'].mean().round(2) > df_intervalo_anos_preenchidos['2016'].mean().round(2)):
       
    anos = pd.Series([2010, 2011, 2012, 2013, 2014, 2015])
    anos.name = 'Year'
    anos.to_csv('../data/lista_anos_aids.csv', index = False)
    print('O intervalo de anos será até 2015')
    
else:
    anos = pd.Series([2010, 2011, 2012, 2013, 2014, 2015, 2016])
    anos.name = 'Year'
    anos.to_csv('../data/lista_anos_aids.csv', index = False)
    print('O intervalo de anos será até 2016')
