==> notebooks/3_extracao_transformacao_massa_aids.ipynb <==
# Etapa 3: Extração e transformação para alcançar a massa de dados com indicadores relevantes no período de tempo entre 2010 a 2016 do dataset "Epidemic transition metrics_Trend of new HIV infections.csv".

## 1. Importando libs

!pip3 install pandas
!pip3 install numpy

import pandas as pd
import numpy as np

## 2. Carregando o dataframe com os dados de todos os países e da faixa temporal.

df_aids_todos_paises = pd.read_csv('../data/Epidemic transition metrics_Trend of new HIV infections.csv')

if df_aids_todos_paises.empty:
    exit("Erro ao ler o arquivo 'df_aids_todos_paises.csv' na pasta 'data'.")
else:
    print("Arquivo 'df_aids_todos_paises.csv' carregado com sucesso.")

## 3. Carregando os países definidos em "lista_paises_aids.csv".

df_lista_paises = pd.read_csv('../data/lista_paises_aids.csv')

df_lista_paises.head()

## 4. Extraindo e transformando dados referentes aos países previamente definidos para "massa_bruta_aids.csv".

#### Seleciona apenas os registros dos países presentes em "lista_paises_aids.csv".

df_paises_filtrados = df_aids_todos_paises.loc[df_aids_todos_paises['Country'].isin(df_lista_paises['Country Name'])]

df_paises_filtrados.head(10)

#### Recupera os anos a serem utilizados para selecionar a faixa temporal da massa final.

df_lista_anos = pd.read_csv('../data/lista_anos_aids.csv')

lista_anos = df_lista_anos['Year'].tolist()

lista_anos

#### Aplica as restrições de anos no dataset original e gera um dataframe com apenas as colunas de 2010 a 2016 e o nome do país.

nomes_colunas_paises_filtrados = list(df_paises_filtrados.columns.values)

valores = {}
for nome_coluna, linha in df_paises_filtrados.items():

    if(nome_coluna == 'Country'):
        valores[nome_coluna] = linha
        continue
        
    valores_linha = []
    for index, item in linha.items():

        valor = item.strip(' ')
        if(valor.isdigit() == False):
            if ('<' in valor):
                valor = valor.strip('<')
                valor = int(valor) - 1

            elif('>' in valor):
                valor = valor.strip('>')
                valor = int(valor) + 1

            elif('...' in valor):
                valor = 0

            else:
                valor = valor.replace(' ', '')

        valores_linha.append(valor)
    valores[nome_coluna] = valores_linha

df_resultado_conversao_valores = pd.DataFrame(valores)

df_resultado_conversao_valores.head()

#### Seleciona as colunas dentro do intervalo definido em "df_lista_anos". As colunas escolhidas terão o ano no seu conteúdo, por exemplo: "2010", "2010_lower" e "2010_upper" referenciam 2010. Logo, seus valores serão convertidos para número e somados para formar uma coluna única, como, por exemplo: "2010".

df_resultado_transformacao = pd.DataFrame()

df_resultado_transformacao['Country'] = df_resultado_conversao_valores['Country']

df_soma_2010 = pd.DataFrame()
df_soma_2010['2010'] = df_resultado_conversao_valores['2010']
df_soma_2010['2010_lower'] = df_resultado_conversao_valores['2010_lower']
df_soma_2010['2010_upper'] = df_resultado_conversao_valores[' 2010_upper']
df_soma_2010 = df_soma_2010.apply(pd.to_numeric, errors='coerce')
df_resultado_transformacao['2010'] = df_soma_2010.sum(axis=1)

df_soma_2011 = pd.DataFrame()
df_soma_2011['2011'] = df_resultado_conversao_valores[' 2011']
df_soma_2011['2011_lower'] = df_resultado_conversao_valores['2011_lower']
df_soma_2011['2011_upper'] = df_resultado_conversao_valores[' 2011_upper']
df_soma_2011 = df_soma_2011.apply(pd.to_numeric, errors='coerce')
df_resultado_transformacao['2011'] = df_soma_2011.sum(axis=1)

df_soma_2012 = pd.DataFrame()
df_soma_2012['2012'] = df_resultado_conversao_valores[' 2012']
df_soma_2012['2012_lower'] = df_resultado_conversao_valores['2012_lower']
df_soma_2012['2012_upper'] = df_resultado_conversao_valores[' 2012_upper']
df_soma_2012 = df_soma_2012.apply(pd.to_numeric, errors='coerce')
df_resultado_transformacao['2012'] = df_soma_2012.sum(axis=1)

df_soma_2013 = pd.DataFrame()
df_soma_2013['2013'] = df_resultado_conversao_valores[' 2013']
df_soma_2013['2013_lower'] = df_resultado_conversao_valores['2013_lower']
df_soma_2013['2013_upper'] = df_resultado_conversao_valores[' 2013_upper']
df_soma_2013 = df_soma_2013.apply(pd.to_numeric, errors='coerce')
df_resultado_transformacao['2013'] = df_soma_2013.sum(axis=1)

df_soma_2014 = pd.DataFrame()
df_soma_2014['2014'] = df_resultado_conversao_valores[' 2014']
df_soma_2014['2014_lower'] = df_resultado_conversao_valores['2014_lower']
df_soma_2014['2014_upper'] = df_resultado_conversao_valores[' 2014_upper']
df_soma_2014 = df_soma_2014.apply(pd.to_numeric, errors='coerce')
df_resultado_transformacao['2014'] = df_soma_2014.sum(axis=1)

df_soma_2015 = pd.DataFrame()
df_soma_2015['2015'] = df_resultado_conversao_valores[' 2015']
df_soma_2015['2015_lower'] = df_resultado_conversao_valores['2015_lower']
df_soma_2015['2015_upper'] = df_resultado_conversao_valores[' 2015_upper']
df_soma_2015 = df_soma_2015.apply(pd.to_numeric, errors='coerce')
df_resultado_transformacao['2015'] = df_soma_2015.sum(axis=1)

df_soma_2016 = pd.DataFrame()
df_soma_2016['2016'] = df_resultado_conversao_valores[' 2016']
df_soma_2016['2016_lower'] = df_resultado_conversao_valores['2016_lower']
df_soma_2016['2016_upper'] = df_resultado_conversao_valores[' 2016_upper']
df_soma_2016 = df_soma_2016.apply(pd.to_numeric, errors='coerce')
df_resultado_transformacao['2016'] = df_soma_2016.sum(axis=1)

df_resultado_transformacao.head()

#### Exporta o conjunto de dados transformado para "massa_bruta_aids.csv".

df_resultado_transformacao.to_csv('../data/massa_bruta_aids.csv', index = False)
