==> notebooks/5_1_classificacao_aids_por_indicadores_socioeconomicos_brasil.ipynb <==
# Etapa 5.1: testar a eficiência de predição dos modelos Dummy, SVC, Guassian Naive Bayes e Árvore de Decisão para os dados do Brasil.

#### O objetivo é testar a acurácia da previsão dos modelos treinados de acordo com os dados do Brasil.

## 1. Instalando e importando as bibliotecas utlizadas nesse notebook.

!pip3 install pandas
!pip3 install numpy
!pip3 install sklearn
!pip3 install graphviz
#!apt-get install graphviz

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.svm import SVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.dummy import DummyClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix, f1_score
from sklearn.preprocessing import StandardScaler
import numpy as np
from sklearn.tree import export_graphviz
import graphviz

## 2. Carga dos dados do Brasil e EUA resultantes das etapas de extração e transformação.

df_indicadores_em_colunas = pd.read_csv('../data/final_transformacao_dados_unificados_aids_indicadores.csv')

df_indicadores_em_colunas.head()

#### Seleção de dados do Brasil.

df_brasil_eua = pd.read_csv('../data/final_transformacao_dados_unificados_aids_indicadores_br_eua.csv')

df_br = df_brasil_eua.iloc[0:6]

df_br.head(6)

## 3. Aplicação dos classificadores Dummy, SVC, Gaussian Naive Bayes e Decision Tree Classifier.

#### A função "classificar" objetiva imprimir os resultados de um modelo de classificação (matriz de confusão e o relatório da classificação).

def classificar(modelo, treino_x, teste_x, treino_y, teste_y, nome_modelo):
    modelo.fit(treino_x, treino_y)
    print("*********  " + nome_modelo + "  *********")
    print("Acurácia treinamento: %.2f%%" % (modelo.score(treino_x, treino_y)*100))
    previsoes = modelo.predict(teste_x)
    print("Acurácia previsão: %.2f%%" % (accuracy_score(teste_y, previsoes)*100))
    print("")
    print("<< Matriz de confusão >>")
    matriz_confusao = confusion_matrix(teste_y, previsoes)
    print(pd.crosstab(teste_y, previsoes, rownames=['Real'], colnames=['Predito'], margins=True))
    print("")
    print("<< Relatório de classificação >>")
    print(classification_report(teste_y, previsoes))

#### Aplicação do "Dummy Classifier".

x = df_indicadores_em_colunas[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

y = df_indicadores_em_colunas['AIDS']


teste_x_brasil = df_br[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

teste_y_brasil = df_br['AIDS']

SEED = 5
np.random.seed(SEED)

classificar(DummyClassifier(strategy='stratified'), x, teste_x_brasil, y, teste_y_brasil, "Dummy")

#### Aplicação do "SVC".

x = df_indicadores_em_colunas[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

y = df_indicadores_em_colunas['AIDS']


teste_x_brasil = df_br[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

teste_y_brasil = df_br['AIDS']

#Define a ordem dos números aleatórios e retira a aleatoriedade da separação dos dados de treino e teste.
SEED = 5
np.random.seed(SEED)

#Reescalando os dados para uma faixa similar a todos.
scaler = StandardScaler()
scaler.fit(x)
treino_x = scaler.transform(x)
teste_x = scaler.transform(teste_x_brasil)

classificar(SVC(), x, teste_x_brasil, y, teste_y_brasil, "SVC")

#### Aplicação do "Gaussian Naive Bayes".

x = df_indicadores_em_colunas[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

y = df_indicadores_em_colunas['AIDS']


teste_x_brasil = df_br[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

teste_y_brasil = df_br['AIDS']

#Define a ordem dos números aleatórios e retira a aleatoriedade da separação dos dados de treino e teste.
SEED = 20
np.random.seed(SEED)


classificar(GaussianNB(), x, teste_x_brasil, y, teste_y_brasil, "Gaussian Naive Bayes")

#### Aplicação do "Decision Tree Classifier". Além disso, é exibido um gráfico para a análise das decisões tomadas pelo algoritmo.

x = df_indicadores_em_colunas[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

y = df_indicadores_em_colunas['AIDS']


teste_x_brasil = df_br[['PIB', 'Usuários Internet', 'Crescimento Populacional', 
                               'Desemprego', 'Investimentos Educação', 'Analfabetismo 15 e 24', 
                               'Analfabetismo 25 e 64', 'Qtd Professores EF', 'Crianças Fora Escola',
                               'Qtd Alunos por professor', 'Qtd Alunos EF']]

teste_y_brasil = df_br['AIDS']

#Define a ordem dos números aleatórios e retira a aleatoriedade da separação dos dados de treino e teste.
SEED = 20
np.random.seed(SEED)

classificar(DecisionTreeClassifier(max_depth=12), x, teste_x_brasil, y, teste_y_brasil, "DecisionTreeClassifier")
